fix rust shit

+++ server/src/world.rs
@@ -11,7 +11,8 @@ use uuid::Uuid;
 use crate::math::Vec2;
 
 // TODO(mluogh): replace with config.toml
-const TICKS_PER_SECOND: i32 = 10;
+const TICKS_PER_SECOND: i8 = 10;
+//#[allow(clippy::cast_precision_loss)]
 const TIME_PER_TICK: f32 = 1.0 / TICKS_PER_SECOND as f32;
 const TIME_PER_TICK_SQUARED: f32 = TIME_PER_TICK * TIME_PER_TICK;
